+++ packages/components-docs/src/CustomDev/index.tsx
@@ -1,4 +1,5 @@
 import AccountManage from './AccountManage';
+import MetadataFetcher from './MetadataFetcher';
 
 
-export default AccountManage;
+export  default MetadataFetcher;