+++ packages/dapp-utils/src/address/index.ts
@@ -1,3 +1,20 @@
 export * from './convertAddress';
 export * from './shortenAddress';
 export * from './validateAddress';
+export type * from './types';
+
+export { checksumHexAddress } from './checksumHexAddress';
+
+export { isBase32Address } from './isBase32Address';
+
+export { isCoreHexAddress } from './isCoreHexAddress';
+
+export { isHexAddress } from './isHexAddress';
+
+export { isPosAddress } from './isPosAddress';
+
+export { isSimplyBase32Address } from './isSimplyBase32Address';
+
+export { isValidNetworkId } from './isValidNetworkId';
+
+export { getCoreHexAddressType, InvalidCoreHexAddressError } from './getCoreHexAddressType';