+++ packages/dapp-utils/src/address/isValidNetworkId.ts
@@ -1,0 +1,9 @@
+export const NET_ID_LIMIT = 0xffffffff;
+
+export function isValidNetworkId(networkId: string | number) {
+  if (typeof networkId === 'string') {
+    return /^([1-9]\d*)$/.test(networkId) && Number(networkId) <= NET_ID_LIMIT;
+  } else {
+    return networkId > 0 && networkId < NET_ID_LIMIT;
+  }
+}